+++ app/lib/humane_transport_app.dart
@@ -1,11 +1,11 @@
+import 'package:app/common/style/style.dart';
 import 'package:app/screens/home/home.dart';
-import 'package:app/style/style.dart';
 import 'package:flutter/material.dart';
 
 class HumaneTransportApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: HomePage(),
+      home: HomeScreen(),
       // This is one way to consume styling properties inside style.dart
       theme: ThemeData(
         textTheme: TextTheme(
